Fix unstable ESRI test
One of the tests was failing after the ESRI funcitonality was merged to
master:

FAIL: Tests loading invalid geometry type
----------------------------------------------------------------------
Traceback (most recent call last):
  File "/path/to/geomet/geomet/tests/esri_test.py", line 161, in test_loads_unsupported_geom_type
    ar.exception))
AssertionError: "Invalid EsriJSON: {'Tetrahedron': [[1, 2, 34], [2, 3, 4], [4, 5, 6]], 'spatialReference': {'wkid': 4326}}" != "Invalid EsriJSON: {'spatialReference': {'wkid': 4326}, 'Tetrahedron': [[1, 2, 34], [2, 3, 4], [4, 5, 6]]}"

This revealed a bug in `geomet.esri.loads` which was allowing
non-string input (such as a dict) to be passed. This should not be
allowed, according to the docstring (which says the method accepts a
string).

This fix was easy: just add a handful of `json.dumps` calls to wrap test
inputs.