Testing: Migrate to pytest, enforce baseline test coverage
pytest is a more mature testing framework and a defacto standard for
Python projects. It also works better for running tests, capturing
coverage stats, and enforcing coverage.

Test coverage is currently about 93%. This isn't great, and we certainly
don't want it to get any worse. Set the baseline now, and it can be
improved in future patches.